+++ runtime/src/main/java/com/instana/android/core/util/ConstantsAndUtil.kt
@@ -23,6 +23,8 @@ import com.instana.android.core.event.models.EffectiveConnectionType
 import com.instana.android.core.event.models.Platform
 import com.instana.android.instrumentation.HTTPCaptureConfig
 import okhttp3.OkHttpClient
+import java.math.BigDecimal
+import java.math.RoundingMode
 import java.net.MalformedURLException
 import java.net.URL
 
@@ -270,4 +272,5 @@ object ConstantsAndUtil {
     fun Int.toDaysInMillis(): Long {
         return maxOf(this, 0) * 24 * 60 * 60 * 1000L
     }
+
 }