+++ runtime/src/main/java/com/instana/android/performance/anr/ANRMonitor.kt
@@ -43,7 +43,8 @@ class ANRMonitor(
             ),
             viewName = Instana.view,
             backendTracingID = null,
-            error = null
+            error = null,
+            customMetric = null
         )
     }
 }